Update README.md
Hi, I'm Prashast, a computer science student with a passion for software engineering. I'm committed to continuous learning, and my love for collaboration and teamwork helps me to develop my skills and deliver innovative solutions to real-world problems. I'm also an excellent communicator, which allows me to work effectively with diverse teams.

When I'm not coding, I enjoy exploring the great outdoors as well as the diverse indoors by immersing myself in a good book. But what really drives me getting to explore anything and everything new coming to the world of software industry.

I'm excited to see where my career takes me and get the chance to know lots and lots of new things.